+++ examples/auth/app/pages/index.tsx
@@ -19,9 +19,9 @@ import {Routes} from ".blitz"
 
 const CurrentUserInfo = () => {
   const session = useSession()
-  // const [currentUser] = useQuery(getUser, {where: {id: session.userId!}})
+  const [currentUser] = useQuery(getUser, {where: {id: session.userId!}})
 
-  return <pre>{JSON.stringify(session, null, 2)}</pre>
+  return <pre>{JSON.stringify(currentUser, null, 2)}</pre>
 }
 
 const UserStuff = () => {
